requirements.txt
Django==1.7.1
Pillow==2.6.1
django-allauth==0.18.0
django-appconf==0.6
django-extensions==1.4.6
django-filter==0.8
django-imagekit==3.2.4
django-taggit==0.12.2
djangorestframework==3.0.0
oauthlib==0.7.2
pilkit==1.1.12
psycopg2==2.5.4
python-openid==2.2.5
requests==2.4.3
requests-oauthlib==0.4.2
shortuuid==0.4.2
six==1.8.0
wsgiref==0.1.2
